add Ratting and change to singleton

+++ src/main/java/com/ooops/lms/database/dao/BookDAO.java
@@ -443,7 +443,7 @@ public class BookDAO implements DatabaseQuery<Book> {
 
                 try (ResultSet resultSet = preparedStatement.executeQuery()) {
                     while (resultSet.next()) {
-                        bookList.add(find(resultSet.getInt("ISBN")));
+                        bookList.add(find(resultSet.getLong("ISBN")));
                     }
                     bookCache.put(keywords, bookList);
                 }
